pyproject.toml
[project]
name = "voice-assistant"
version = "2.7.0"
description = "智能语音助手 - Agent Skills + MCP + Pipecat (KWS + ASR + MCP + Skills)"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
keywords = ["voice", "assistant", "speech-recognition", "wake-word", "ai", "mcp", "react-agent", "agent-skills", "anthropic"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]

dependencies = [
    "sherpa-onnx>=1.12.19",
    "numpy>=1.24.0",
    "pyaudio>=0.2.11",
    "requests>=2.32.5",
    "httpx>=0.27.0",  # 异步 HTTP 客户端（用于 Vision API）
    "aiofiles>=24.1.0",  # 异步文件 I/O（用于 Vision 图片读取）
    "dashscope>=1.25.4",
    "pillow>=11.3.0",
    "soundcard>=0.4.2",
    "python-dotenv>=1.0.0",  # 环境变量管理
    "python-frontmatter>=1.0.0",  # 解析 SKILL.md YAML frontmatter
    # MCP 依赖
    "mcp>=1.0.0", # 官方 MCP Python SDK
    "realtimetts>=0.5.7",
    "pyttsx3>=2.99",
    "piper-tts>=1.3.0",
    "pipecat-ai[local,moondream]>=0.0.98",  # 解析 SKILL.md YAML frontmatter
    "websocket>=0.2.1",
    "websockets>=16.0",
    # "pydantic-settings[dotenv]>=2.12.0",  # ❌ 不需要，python-dotenv 已经足够
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "aiohttp>=3.13.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
voice-assistant = "main:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"
